@@ -68,17 +68,18 @@
     "i18n": "ng-xi18n"
   },
   "dependencies": {
-    "@angular/common": "2.4.6",
-    "@angular/compiler": "2.4.6",
-    "@angular/core": "2.4.6",
-    "@angular/forms": "2.4.6",
-    "@angular/http": "2.4.6",
-    "@angular/material": "^2.0.0-beta.2",
-    "@angular/platform-browser": "2.4.6",
-    "@angular/platform-browser-dynamic": "2.4.6",
-    "@angular/platform-server": "2.4.6",
-    "@angular/router": "3.4.6",
-    "@angularclass/conventions-loader": "^1.0.2",
+    "@angular/animations": "^4.0.2",
+    "@angular/common": "^4.0.2",
+    "@angular/compiler": "^4.0.2",
+    "@angular/core": "^4.0.2",
+    "@angular/forms": "^4.0.2",
+    "@angular/http": "^4.0.2",
+    "@angular/material": "2.0.0-beta.2",
+    "@angular/platform-browser": "^4.0.0",
+    "@angular/platform-browser-dynamic": "^4.0.2",
+    "@angular/platform-server": "^4.0.2",
+    "@angular/router": "^4.0.2",
+    "@angularclass/conventions-loader": "^1.0.13",
     "@angularclass/hmr": "~1.2.2",
     "@angularclass/hmr-loader": "~3.0.2",
     "bootstrap": "^4.0.0-alpha.6",
@@ -89,10 +90,10 @@
     "jasmine-core": "^2.5.2",
     "reflect-metadata": "^0.1.9",
     "rxjs": "5.0.2",
-    "zone.js": "0.7.6"
+    "zone.js": "0.8.4"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "2.4.6",
+    "@angular/compiler-cli": "4.0.2",
     "@types/hammerjs": "^2.0.33",
     "@types/jasmine": "^2.2.34",
     "@types/node": "^7.0.0",
