@@ -2,7 +2,10 @@ import { Component,
 		 ChangeDetectorRef,
 		 ChangeDetectionStrategy,
 		 OnInit,
+		 OnDestroy,
 		 ViewEncapsulation } from '@angular/core';
+
+import { Subscription } from 'rxjs/Subscription';
 
 import { LoaderBlockService } from './loader-block.service';
 
@@ -13,15 +16,20 @@ import { LoaderBlockService } from './loader-block.service';
 	encapsulation: ViewEncapsulation.None,
 	changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class LoaderBlockComponent implements OnInit {
+export class LoaderBlockComponent implements OnInit, OnDestroy {
 	loading: boolean;
+	loadingSub: Subscription;
 
 	constructor (private cd: ChangeDetectorRef, private loaderBlockService: LoaderBlockService) {}
 
 	ngOnInit(): void {
-		this.loaderBlockService.show.subscribe((show: boolean) => {
+		this.loadingSub = this.loaderBlockService.show.subscribe((show: boolean) => {
 			this.loading = show;
 			this.cd.markForCheck();
 		});
+	}
+
+	ngOnDestroy(): void {
+		this.loadingSub.unsubscribe();
 	}
 }
