@@ -3,7 +3,10 @@ import { Pipe, PipeTransform } from '@angular/core';
 
 import { Course } from '../interfaces';
 
-@Pipe({name: 'orderBy'})
+@Pipe({
+	name: 'orderBy',
+	pure: false
+})
 export class OrderByPipe implements PipeTransform {
 	transform(array: Course[], fields: string[], order: string[]): Course[] {
 		return _.orderBy(array, fields, order);
