@@ -20,18 +20,23 @@ import { FilerByNamePipe } from '../../../common/pipes/filter-by-name.pipe';
 	templateUrl: './find-course.template.html',
 	changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class FindCourseComponent {
+export class FindCourseComponent implements OnInit {
 	@Input() courses: Course[];
 	@Output() coursesChange = new EventEmitter();
 
 	query: string;
+	private unsortedCourses: Course[];
 
 	constructor(private filterPipe: FilerByNamePipe) {
 		this.query = '';
 	}
 
+	ngOnInit(): void {
+		this.unsortedCourses = this.courses;
+	}
+
 	findCourse(): void {
-		let filteredArr = this.filterPipe.transform(this.courses, this.query);
+		let filteredArr = this.filterPipe.transform(this.unsortedCourses, this.query);
 
 		if (filteredArr.length) {
 			this.coursesChange.emit(filteredArr);
